Вынесла фукнции из main в отдельные экспортируемые модули
