+++ src/view/board.js
@@ -1,0 +1,5 @@
+export const createBoardTemplate = () =>
+  `<section class="board container">
+    <div class="board__tasks"></div>
+  </section>`
+;
