+++ src/view/load-more-btn.js
@@ -1,0 +1,3 @@
+export const createLoadMoreBtnTemplate = () =>
+  `<button class="load-more" type="button">load more</button>`
+;
